refactor: Organiza os métodos nas classes criadas no pacote model (FileHandler e VideoManager). Stayle: Retirado alguns imports não usados, espaços a mais. Instancias não usadas.
